Update add_sim_index_number.R
Variable "idx" containes indices of the first row of the dataset, the last row of each simulation subproblem (each trial), and the last row of the dataset + 1. More specific, the first two values of idx represent the first simulation subproblem, then, 3rd value of idx represents the last row of the 2nd simulation subproblem, etc.
In the "original" version For loop assigns simulation index "i" to the rows from idx[i] to the idx[i+1]-1. I believe that this is wrong since idx[i], for i > 1, represents the last row of the previous simulation (i-1 simulation). Therefore, I have corrected the For loop.